requirements.txt
beautifulsoup4==4.9.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
matplotlib==3.3.2
numpy==1.18.5
pandas==1.1.2
pandas-datareader==0.9.0
Pillow==7.2.0
scikit-learn==0.23.2
sklearn==0.0
soupsieve==2.0.1
streamlit==0.67.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0

